+++ packages/react-devtools-shared/src/devtools/views/Settings/GeneralSettings.js
@@ -11,7 +11,7 @@ import * as React from 'react';
 import {useContext, useMemo} from 'react';
 import {SettingsContext} from './SettingsContext';
 import {StoreContext} from '../context';
-import {CHANGE_LOG_URL} from 'react-devtools-shared/src/constants';
+import {CHANGE_LOG_URL} from 'react-devtools-shared/src/devtools/constants';
 
 import styles from './SettingsShared.css';
