+++ packages/react-devtools-shared/src/utils.js
@@ -53,9 +53,8 @@ import {localStorageGetItem, localStorageSetItem} from './storage';
 import {meta} from './hydration';
 import isArray from './isArray';
 
-import type {ComponentFilter, ElementType} from './types';
+import type {ComponentFilter, ElementType, BrowserTheme} from './types';
 import type {LRUCache} from 'react-devtools-shared/src/types';
-import type {BrowserTheme} from 'react-devtools-shared/src/devtools/views/DevTools';
 
 // $FlowFixMe[method-unbinding]
 const hasOwnProperty = Object.prototype.hasOwnProperty;
